+++ mangadex-dl.py
@@ -16,71 +16,10 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-import argparse, os
+import argparse
 from mangadex_dl_functions import *
 
 MANGADEX_DL_VERSION = "1.0"
-
-def dl(manga_url, args):
-	manga_uuid = get_uuid(manga_url)
-	manga_title, manga_title_en = get_title(manga_uuid, args.language)
-	
-	print("\n[{:2}/{:2}] TITLE: {}\n".format(args.manga_urls.index(manga_url)+1, len(args.manga_urls), manga_title))
-	
-	# check available chapters
-	chapters_info = get_chapters_info(manga_uuid, args.language)
-	
-	if chapters_info["total"] == 0:
-		raise ValueError("No chapters available to download!")
-	
-	chapters_list = get_chapters_list(manga_uuid, chapters_info["total"], args.language)
-	
-	# duplicate check
-	if args.resolve != "all":
-		duplicated_chapters_list = get_duplicated_chapters(chapters_list)
-		if len(duplicated_chapters_list) != 0:
-			chapters_list = resolve_duplicated_chapters(chapters_list, duplicated_chapters_list, args.resolve)
-	
-	# print chapters list
-	print_available_chapters(chapters_list)
-	
-	# i/o for chapters to download
-	if args.download == None:
-		dl_input = input("\nEnter chapter(s) to download:"\
-				 "\n(see README for examples of valid format)"\
-				 "\n> ")
-	else:
-		dl_input = args.download
-	
-	dl_list = parse_range(dl_input)
-	
-	# requested chapters list in dl_range
-	requested_chapters = get_requested_chapters(chapters_list, dl_list)
-	if len(requested_chapters) == 0:
-		raise ValueError("Empty list of chapters. Make sure you enter the correct download range!")
-	
-	# download images
-	if os.path.isdir(args.outdir):
-		out_directory = args.outdir
-	else:
-		out_directory = "."
-	
-	manga_directory = os.path.join(out_directory, manga_title_en)
-	if not os.path.exists(manga_directory):
-		try:
-			os.makedirs(manga_directory)
-		except OSError:
-			manga_directory = os.path.join(out_directory, "Manga {}".format(manga_uuid))
-			if not os.path.exists(manga_directory):
-				os.makedirs(manga_directory)
-	
-	download_chapters(requested_chapters, manga_directory, args.datasaver)
-	
-	# archive
-	if args.archive != None:
-		archive_manga_directory(manga_directory, out_directory, args.archive, args.keep)
-	
-	print("\nManga \"{}\" was successfully downloaded".format(manga_title))
 
 def initialize():
 	help_str = "Examples of valid input for a range of downloadable chapters: "\
